python-services/requirements.txt
# Python services requirements for enhanced game theory platform

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Game theory and computation
nashpy==0.0.24
nashpy[scipy]==0.0.24
nashpy[numpy]==0.0.24
gambit==16.0.1
pygambit==16.0.1

# Machine learning and AI
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4
scikit-learn==1.3.2
torch==2.1.2
torchvision==0.16.2
stable-baselines3==2.2.1
pettingzoo==1.24.3

# Data processing
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# API and web services
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Database and storage
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
jinja2==3.1.2
pyyaml==6.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-mermaid2-plugin==1.1.1

# Monitoring and logging
structlog==23.2.0
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Optional: Additional ML libraries
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.3

# Optional: Advanced computation
sympy==1.12
networkx==3.2.1
graphviz==0.20.1

# Optional: Geopolitical data sources
geopandas==0.14.1
shapely==2.0.2
folium==0.15.0
